[rename] tts 폴더 내 table 폴더 생성 #84
