+++ src/components/tts/TTSMainContent.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react';
 
 import { Button } from '@/components/buttons/CommonButton';
 import { SaveButton, UploadButton } from '@/components/buttons/IconButton';
-import { TTSTable } from '@/components/tts/TTSTable';
+import { TTSTable } from '@/components/tts/table/TTSTable';
 
 const TTSMainContent: React.FC = () => {
   const [items, setItems] = useState([
