+++ src/components/tts/table/TTSTable.tsx
@@ -4,8 +4,8 @@ import { TbCirclePlus, TbDownload, TbReload, TbTrash } from 'react-icons/tb';
 
 import { IconButton } from '@/components/buttons/IconButton';
 import ViewButtonGroup from '@/components/buttons/ViewFilterButton';
-import { TTSTableGrid } from '@/components/tts/TTSTableGrid';
-import { TtsTableList } from '@/components/tts/TTSTableList';
+import { TTSTableGrid } from '@/components/tts/table/TTSTableGrid';
+import { TtsTableList } from '@/components/tts/table/TTSTableList';
 import { Checkbox } from '@/components/ui/checkbox';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Separator } from '@/components/ui/separator';
